+++ PPADisenioCore/src/main/java/edu/core/entities/Pais.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Pais {
-	private Long idPais;
+	private int idPais;
 	private String nombre;
 	private List<Provincia> provincias = new ArrayList<Provincia>();
 	
@@ -14,10 +14,10 @@ public class Pais {
 	public void setProvincias(List<Provincia> provincias) {
 		this.provincias = provincias;
 	}
-	public Long getIdPais() {
+	public int getIdPais() {
 		return idPais;
 	}
-	public void setIdPais(Long idPais) {
+	public void setIdPais(int idPais) {
 		this.idPais = idPais;
 	}
 	public String getNombre() {
