+++ PPADisenioCore/src/main/java/edu/core/entities/Provincia.java
@@ -1,7 +1,7 @@
 package edu.core.entities;
 
 public class Provincia {
-	private Long idProvincia ;
+	private int idProvincia ;
 	private String nombre ;
 	private Pais pais;
 
@@ -11,10 +11,10 @@ public class Provincia {
 	public void setPais(Pais pais) {
 		this.pais = pais;
 	}
-	public Long getIdProvincia() {
+	public int getIdProvincia() {
 		return idProvincia;
 	}
-	public void setIdProvincia(Long idProvincia) {
+	public void setIdProvincia(int idProvincia) {
 		this.idProvincia = idProvincia;
 	}
 	public String getNombre() {
